backend/requirements.txt
# FastAPI Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# OCR & Image Processing
Pillow==11.1.0
pdf2image==1.17.0
opencv-python-headless==4.10.0.84
numpy>=2.0.0,<2.3.0
azure-cognitiveservices-vision-computervision==0.9.0
requests==2.32.3
pykakasi==2.3.0

# Excel/CSV Processing
openpyxl==3.1.5
pandas==2.2.3
xlrd==2.0.1

# PDF Processing
PyPDF2==3.0.1
reportlab==4.2.5

# Email
python-dotenv==1.0.1
aiosmtplib==3.0.2
jinja2==3.1.4

# Validation
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.2.0

# Date/Time
python-dateutil==2.9.0.post0
pytz==2024.2

# HTTP Requests
httpx==0.28.0
aiohttp==3.11.11

# Utilities
python-slugify==8.0.4
qrcode[pil]==8.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# CORS
fastapi-cors==0.0.6

# Redis (opcional)
redis==5.2.1

# Face Detection Enhancement
mediapipe==0.10.14

# OCR Enhancement for Japanese Documents
easyocr==1.7.2

# Logging
loguru==0.7.2

# Rate Limiting
slowapi==0.1.9
